Cleaned up duplicate functions for this fix, made the fix less a patch but a real fix.
