+++ app/dashboard/lib/open-source-filter.ts
@@ -1,9 +1,9 @@
-import partnershipsData from "@/partnerships/partnerships.json";
 import type {
   BoardTimeFilter,
   OpenSourceColumnId,
   OpenSourceEntry,
 } from "../components/types";
+import { buildPartnershipNameMatchSet, normalizePartnerName } from "./partnership-name-match";
 
 type CompletedPartnershipLite = {
   partnershipId?: number;
@@ -25,29 +25,6 @@ type Params = {
   fullPartnerships: NameOption[];
   isWithinCurrentMonth: (value?: string | null) => boolean;
 };
-
-function normalizePartnerName(s: string | null | undefined): string {
-  return (s ?? "").trim().toLowerCase();
-}
-
-function buildPartnershipNameMatchSet(
-  partnershipId: number | null,
-  displayName: string | null,
-  availablePartnerships: NameOption[],
-  fullPartnerships: NameOption[]
-): Set<string> {
-  const out = new Set<string>();
-  if (displayName) out.add(normalizePartnerName(displayName));
-  if (partnershipId != null) {
-    const fromJson = partnershipsData.partnerships.find((p) => p.id === partnershipId)?.name;
-    if (fromJson) out.add(normalizePartnerName(fromJson));
-    const fromAvail = availablePartnerships.find((p) => p.id === partnershipId)?.name;
-    if (fromAvail) out.add(normalizePartnerName(fromAvail));
-    const fromFull = fullPartnerships.find((p) => p.id === partnershipId)?.name;
-    if (fromFull) out.add(normalizePartnerName(fromFull));
-  }
-  return out;
-}
 
 function openSourceDateForMonthFilter(entry: OpenSourceEntry): string | null {
   return entry.dateModified ?? entry.dateCreated ?? null;
